RavenDB-17055 : fix flaky test
